pyproject.toml
[project]
name = "sokuho"
version = "0.1.0"
description = "SOKUHO PDF to SQL Server ETL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.0",
    "pandas>=2.2.0",
    "pyodbc>=5.1.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["etl"]
